Omit slow tests in start.sh script.
This makes the script skip the js tests and other tests
before running the app. With this commit, when new people
run the app for the first time, it starts the web server
in a more reasonable amount of time.

This improvement inspired by feedback by a tester of our
documentation who remarked that they gave errors related
to not launching the JS-based tests properly.